requirements/production.txt
Django==3.2
django-extensions
django-tinymce
Faker
Pillow
psycopg2-binary
django-import-export
six
Werkzeug
pyOpenSSL
python-social-auth
social-auth-core
redis
djangorestframework
bleach
django-filter
djoser
djangorestframework-simplejwt==4.3.0
social-auth-app-django==4.0.0
django-cors-headers==3.13.0